requirements.txt
yfinance>=0.2.40
feedparser>=6.0.11
requests>=2.32.0
feedgen>=1.0.0
python-dateutil>=2.9.0
piper-tts>=1.2.0; platform_system != "Darwin"
elevenlabs>=1.0.0
python-dotenv>=1.0.0
lxml>=5.0.0
Pillow>=10.0.0
pytest>=8.0.0
PyYAML>=6.0
